Add files via upload
The objective of this project would be to build an SDN controller. You will be having a dumb/simple packet forwarding switch and an intelligent controller. The dumb packet forwarding switch will be talking to the controller to make forwarding decisions. Based on how you program the controller, the packet forwarding switch will either act as a router or a switch. The controller and the packet forwarding switch will interact with each other using the OpenFlow protocol. We will be using Mininet network emulation software to emulate the testbeds and networks to test and validate the controller and the switch interaction.